Revert "Balder/group commits to tls 2 remaining rebased"
