+++ searchlib/src/vespa/searchlib/transactionlog/domain.cpp
@@ -1,8 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "domain.h"
-#include "domainpart.h"
-#include "session.h"
 #include <vespa/vespalib/util/stringfmt.h>
 #include <vespa/vespalib/util/closuretask.h>
 #include <vespa/vespalib/io/fileutil.h>
@@ -17,11 +15,10 @@
 LOG_SETUP(".transactionlog.domain");
 
 using vespalib::string;
-using vespalib::make_string_short::fmt;
+using vespalib::make_string;
 using vespalib::LockGuard;
 using vespalib::makeTask;
 using vespalib::makeClosure;
-using vespalib::makeLambdaTask;
 using vespalib::Monitor;
 using vespalib::MonitorGuard;
 using search::common::FileHeaderContext;
@@ -32,10 +29,17 @@ namespace search::transactionlog {
 
 VESPA_THREAD_STACK_TAG(domain_commit_executor);
 
+DomainConfig::DomainConfig()
+    : _encoding(Encoding::Crc::xxh64, Encoding::Compression::none),
+      _compressionLevel(9),
+      _partSizeLimit(0x10000000), // 256M
+      _chunkSizeLimit(0x40000),   // 256k
+      _chunkAgeLimit(10ms)
+{ }
+
 Domain::Domain(const string &domainName, const string & baseDir, Executor & commitExecutor,
                Executor & sessionExecutor, const DomainConfig & cfg, const FileHeaderContext &fileHeaderContext)
     : _config(cfg),
-      _currentChunk(std::make_unique<Chunk>()),
       _lastSerial(0),
       _singleCommiter(std::make_unique<vespalib::ThreadStackExecutor>(1, 128*1024)),
       _commitExecutor(commitExecutor),
@@ -56,10 +60,10 @@ Domain::Domain(const string &domainName, const string & baseDir, Executor & comm
 {
     int retval(0);
     if ((retval = makeDirectory(_baseDir.c_str())) != 0) {
-        throw runtime_error(fmt("Failed creating basedirectory %s r(%d), e(%d)", _baseDir.c_str(), retval, errno));
+        throw runtime_error(make_string("Failed creating basedirectory %s r(%d), e(%d)", _baseDir.c_str(), retval, errno));
     }
     if ((retval = makeDirectory(dir().c_str())) != 0) {
-        throw runtime_error(fmt("Failed creating domaindir %s r(%d), e(%d)", dir().c_str(), retval, errno));
+        throw runtime_error(make_string("Failed creating domaindir %s r(%d), e(%d)", dir().c_str(), retval, errno));
     }
     SerialNumList partIdVector = scanDir();
     const int64_t lastPart = partIdVector.empty() ? 0 : partIdVector.back();
@@ -123,12 +127,7 @@ private:
     bool              & _pendingSync;
 };
 
-Domain::~Domain() {
-    MonitorGuard guard(_currentChunkMonitor);
-    guard.broadcast();
-    commitChunk(grabCurrentChunk(guard), guard);
-    _singleCommiter->shutdown().sync();
-}
+Domain::~Domain() { }
 
 DomainInfo
 Domain::getDomainInfo() const
@@ -309,89 +308,10 @@ waitPendingSync(vespalib::Monitor &syncMonitor, bool &pendingSync)
 
 }
 
-Domain::Chunk::Chunk()
-    : _data(size_t(-1)),
-      _callBacks(),
-      _firstArrivalTime()
-{}
-
-Domain::Chunk::~Chunk() = default;
-
 void
-Domain::Chunk::add(const Packet &packet, Writer::DoneCallback onDone) {
-    if (_callBacks.empty()) {
-        _firstArrivalTime = vespalib::steady_clock::now();
-    }
-    _data.merge(packet);
-    _callBacks.emplace_back(std::move(onDone));
-}
-
-vespalib::duration
-Domain::Chunk::age() const {
-    if (_callBacks.empty()) {
-        return 0ms;
-    }
-    return (vespalib::steady_clock::now() - _firstArrivalTime);
-}
-
-void
-Domain::commit(const Packet & packet, Writer::DoneCallback onDone) {
-    vespalib::MonitorGuard guard(_currentChunkMonitor);
-    if (_lastSerial >= packet.range().from()) {
-        throw runtime_error(fmt("Incoming serial number(%" PRIu64 ") must be bigger than the last one (%" PRIu64 ").",
-                                packet.range().from(), _lastSerial));
-    } else {
-        _lastSerial = packet.range().to();
-    }
-    _currentChunk->add(packet, std::move(onDone));
-    commitIfFull(guard);
-}
-
-void
-Domain::commitIfFull(const vespalib::MonitorGuard &guard) {
-    if (_currentChunk->sizeBytes() > _config.getChunkSizeLimit()) {
-        auto completed = grabCurrentChunk(guard);
-        if (completed) {
-            commitChunk(std::move(completed), guard);
-        }
-    }
-}
-
-std::unique_ptr<Domain::Chunk>
-Domain::grabCurrentChunk(const vespalib::MonitorGuard & guard) {
-    assert(guard.monitors(_currentChunkMonitor));
-    auto chunk = std::move(_currentChunk);
-    _currentChunk = std::make_unique<Chunk>();
-    return chunk;
-}
-
-void
-Domain::commitIfStale() {
-    vespalib::MonitorGuard guard(_currentChunkMonitor);
-    commitIfStale(guard);
-}
-
-void
-Domain::commitIfStale(const vespalib::MonitorGuard & guard) {
-    assert(guard.monitors(_currentChunkMonitor));
-    if ((_currentChunk->age() > _config.getChunkAgeLimit()) && ! _currentChunk->getPacket().empty()) {
-        commitChunk(grabCurrentChunk(guard), guard);
-    }
-}
-
-void
-Domain::commitChunk(std::unique_ptr<Chunk> chunk, const vespalib::MonitorGuard & chunkOrderGuard) {
-    assert(chunkOrderGuard.monitors(_currentChunkMonitor));
-    if ( ! chunk->getPacket().empty()) {
-        _singleCommiter->execute( makeLambdaTask([this, chunk = std::move(chunk)]() mutable {
-            doCommit(std::move(chunk));
-        }));
-    }
-}
-
-void
-Domain::doCommit(std::unique_ptr<Chunk> chunk) {
-    const Packet & packet = chunk->getPacket();
+Domain::commit(const Packet & packet, Writer::DoneCallback onDone)
+{
+    (void) onDone;
     DomainPart::SP dp(_parts.rbegin()->second);
     vespalib::nbostream_longlivedbuf is(packet.getHandle().data(), packet.getHandle().size());
     Packet::Entry entry;
@@ -411,9 +331,6 @@ Domain::doCommit(std::unique_ptr<Chunk> chunk) {
         vespalib::File::sync(dir());
     }
     dp->commit(entry.serial(), packet);
-    if (_config.getFSyncOnCommit()) {
-        dp->sync();
-    }
     cleanSessions();
 }
 
@@ -439,7 +356,7 @@ Domain::erase(SerialNum to)
 
 int
 Domain::visit(const Domain::SP & domain, SerialNum from, SerialNum to,
-              std::unique_ptr<Destination> dest)
+              std::unique_ptr<Session::Destination> dest)
 {
     assert(this == domain.get());
     cleanSessions();
@@ -525,7 +442,7 @@ Domain::scanDir()
             continue;
         const char *p = ename + wantPrefixLen + 1;
         uint64_t num = strtoull(p, NULL, 10);
-        string checkName = fmt("%s-%016" PRIu64, _name.c_str(), num);
+        string checkName = make_string("%s-%016" PRIu64, _name.c_str(), num);
         if (strcmp(checkName.c_str(), ename) != 0)
             continue;
         res.push_back(static_cast<SerialNum>(num));
