--- searchlib/src/vespa/searchlib/transactionlog/domainconfig.cpp
@@ -1,16 +1,0 @@
-// Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-
-#include "domainconfig.h"
-
-namespace search::transactionlog {
-
-DomainConfig::DomainConfig()
-    : _encoding(Encoding::Crc::xxh64, Encoding::Compression::none),
-      _compressionLevel(9),
-      _fSyncOnCommit(false),
-      _partSizeLimit(0x10000000), // 256M
-      _chunkSizeLimit(0x40000),   // 256k
-      _chunkAgeLimit(10ms)
-{ }
-
-}
