+++ searchlib/src/vespa/searchlib/transactionlog/domainpart.cpp
@@ -252,7 +252,7 @@ DomainPart::buildPacketMapping(bool allowTruncate)
 
 DomainPart::DomainPart(const string & name, const string & baseDir, SerialNum s, Encoding encoding,
                        uint8_t compressionLevel, const FileHeaderContext &fileHeaderContext, bool allowTruncate)
-    : _encoding(encoding),
+    : _encoding(encoding.getCrc(), Encoding::Compression::none), //TODO We do not yet support compression
       _compressionLevel(compressionLevel),
       _lock(),
       _fileLock(),
@@ -429,8 +429,7 @@ DomainPart::commit(SerialNum firstSerial, const Packet &packet)
     }
 }
 
-void
-DomainPart::sync()
+void DomainPart::sync()
 {
     SerialNum syncSerial(0);
     {
