+++ searchlib/src/vespa/searchlib/transactionlog/ichunk.h
@@ -3,6 +3,9 @@
 #pragma once
 
 #include "common.h"
+#include <memory>
+
+namespace vespalib { class nbostream; }
 
 namespace search::transactionlog {
