+++ searchlib/src/vespa/searchlib/transactionlog/translogserver.h
@@ -1,29 +1,30 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #pragma once
 
-#include "domainconfig.h"
+#include "domain.h"
 #include <vespa/vespalib/util/document_runnable.h>
 #include <vespa/vespalib/util/threadstackexecutor.h>
 #include <vespa/document/util/queue.h>
 #include <vespa/fnet/frt/invokable.h>
 #include <mutex>
 
+
 class FRT_Supervisor;
 class FNET_Transport;
-class FNET_Task;
 
 namespace search::common { class FileHeaderContext; }
+
 namespace search::transactionlog {
 
 class TransLogServerExplorer;
-class Domain;
 
 class TransLogServer : public document::Runnable, private FRT_Invokable, public Writer
 {
 public:
     friend class TransLogServerExplorer;
-    using SP = std::shared_ptr<TransLogServer>;
-    using DomainSP = std::shared_ptr<Domain>;
+    typedef std::unique_ptr<TransLogServer> UP;
+    typedef std::shared_ptr<TransLogServer> SP;
+
     TransLogServer(const vespalib::string &name, int listenPort, const vespalib::string &baseDir,
                    const common::FileHeaderContext &fileHeaderContext, const DomainConfig & cfg, size_t maxThreads);
     TransLogServer(const vespalib::string &name, int listenPort, const vespalib::string &baseDir,
@@ -32,7 +33,6 @@ public:
                    const common::FileHeaderContext &fileHeaderContext);
     ~TransLogServer() override;
     DomainStats getDomainStats() const;
-    void commitIfStale();
     void commit(const vespalib::string & domainName, const Packet & packet, DoneCallback done) override;
     TransLogServer & setDomainConfig(const DomainConfig & cfg);
 
@@ -71,13 +71,13 @@ private:
     void downSession(FRT_RPCRequest *req);
 
     std::vector<vespalib::string> getDomainNames();
-    DomainSP findDomain(vespalib::stringref name);
+    Domain::SP findDomain(vespalib::stringref name);
     vespalib::string dir()        const { return _baseDir + "/" + _name; }
     vespalib::string domainList() const { return dir() + "/" + _name + ".domains"; }
 
     static const Session::SP & getSession(FRT_RPCRequest *req);
 
-    using DomainList = std::map<vespalib::string, DomainSP >;
+    using DomainList = std::map<vespalib::string, Domain::SP >;
 
     vespalib::string                    _name;
     vespalib::string                    _baseDir;
@@ -87,15 +87,12 @@ private:
     std::unique_ptr<FastOS_ThreadPool>  _threadPool;
     std::unique_ptr<FNET_Transport>     _transport;
     std::unique_ptr<FRT_Supervisor>     _supervisor;
-    std::unique_ptr<FNET_Task>          _staleCommitTask;
     DomainList                          _domains;
-    mutable std::mutex                  _domainMutex;          // Protects _domains
-    std::condition_variable             _domainCondition;
+    mutable std::mutex                  _lock;          // Protects _domains
     std::mutex                          _fileLock;      // Protects the creating and deleting domains including file system operations.
     document::Queue<FRT_RPCRequest *>   _reqQ;
     const common::FileHeaderContext    &_fileHeaderContext;
     using Guard = std::lock_guard<std::mutex>;
-    using MonitorGuard = std::unique_lock<std::mutex>;
 };
 
 }
